moved utils outside of subdirs
